requirements.txt
amqp==2.2.2
Babel==2.5.1
billiard==3.5.0.3
celery==4.1.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
django-toolbelt==0.0.1
djangorestframework==3.7.1
flower==0.9.2
gunicorn==19.7.1
kombu==4.1.0
psycopg2==2.7.3.2
pytz==2017.2
static3==0.7.0
tornado==4.5.2
vine==1.1.4
